数据对齐：写入当日点赞变更

+++ xhs-data-align/src/main/java/org/n1vnhil/xhs/data/align/constants/TableConstants.java
@@ -4,7 +4,7 @@ public class TableConstants {
 
     private static final String TABLE_NAME_SEPARATE = "_";
 
-    public static String buildTableNameSuffix(String date, int hashKey) {
+    public static String buildTableNameSuffix(String date, Long hashKey) {
         return date + TABLE_NAME_SEPARATE + hashKey;
     }
